pyproject.toml
[project]
name = "Algophilia"
version = "0.1.0"
description = ""
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "dependency-injector>=4.49.0",
    "python-dotenv>=1.2.2",
    "pydantic>=2.13.3",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "flet>=0.84.0",
]
